update logpose ActiveDriversResult model

+++ logpose/reference.go
@@ -4,8 +4,8 @@ type ActiveDriversResult struct {
 	DriverID    string `json:"driver_id"'`
 	DriverName  string `json:"driver_name"'`
 	Coordinates struct {
-		Lat float32 `json:"lat"`
-		Lng float32 `json:"lng"`
+		Lat float64 `json:"lat"`
+		Lng float64 `json:"lng"`
 	} `json:"coordinates"`
 	LastCheckin string `json:"last_checkin"`
 }
